move `StartSessionAction` back to /browser/, add indirection for holdForSpeech-feature
https://github.com/microsoft/vscode-copilot/issues/3857
