+++ src/vs/workbench/contrib/inlineChat/electron-sandbox/inlineChatActions.ts
@@ -2,66 +2,24 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { KeyChord, KeyCode, KeyMod } from 'vs/base/common/keyCodes';
+import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
-import { EditorAction2 } from 'vs/editor/browser/editorExtensions';
 import { ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
-import { InlineChatController, InlineChatRunOptions } from 'vs/workbench/contrib/inlineChat/browser/inlineChatController';
-import { AbstractInlineChatAction } from 'vs/workbench/contrib/inlineChat/browser/inlineChatActions';
-import { LOCALIZED_START_INLINE_CHAT_STRING, START_INLINE_CHAT } from '../browser/inlineChatActions';
+import { InlineChatController } from 'vs/workbench/contrib/inlineChat/browser/inlineChatController';
+import { AbstractInlineChatAction, setHoldForSpeech } from 'vs/workbench/contrib/inlineChat/browser/inlineChatActions';
 import { disposableTimeout } from 'vs/base/common/async';
 import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
 import { ICommandService } from 'vs/platform/commands/common/commands';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { StartVoiceChatAction, StopListeningAction } from 'vs/workbench/contrib/chat/electron-sandbox/actions/voiceChatActions';
 import { IChatExecuteActionContext } from 'vs/workbench/contrib/chat/browser/actions/chatExecuteActions';
-import { CTX_INLINE_CHAT_HAS_PROVIDER, CTX_INLINE_CHAT_VISIBLE, InlineChatConfigKeys } from 'vs/workbench/contrib/inlineChat/common/inlineChat';
-import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { CTX_INLINE_CHAT_VISIBLE, InlineChatConfigKeys } from 'vs/workbench/contrib/inlineChat/common/inlineChat';
 import { HasSpeechProvider, ISpeechService } from 'vs/workbench/contrib/speech/common/speechService';
 import { localize2 } from 'vs/nls';
-
-
-export class StartSessionAction extends EditorAction2 {
-
-	constructor() {
-		super({
-			id: 'inlineChat.start',
-			title: { value: LOCALIZED_START_INLINE_CHAT_STRING, original: 'Start Inline Chat' },
-			category: AbstractInlineChatAction.category,
-			f1: true,
-			precondition: ContextKeyExpr.and(CTX_INLINE_CHAT_HAS_PROVIDER, EditorContextKeys.writable),
-			keybinding: {
-				when: EditorContextKeys.focus,
-				weight: KeybindingWeight.WorkbenchContrib,
-				primary: KeyMod.CtrlCmd | KeyCode.KeyI,
-				secondary: [KeyChord(KeyMod.CtrlCmd | KeyCode.KeyK, KeyCode.KeyI)],
-			},
-			icon: START_INLINE_CHAT
-		});
-	}
-
-
-	override runEditorCommand(accessor: ServicesAccessor, editor: ICodeEditor, ..._args: any[]) {
-
-		const configService = accessor.get(IConfigurationService);
-		const speechService = accessor.get(ISpeechService);
-
-		if (configService.getValue<boolean>(InlineChatConfigKeys.HoldToSpeech) // enabled
-			&& speechService.hasSpeechProvider // possible
-		) {
-			holdForSpeech(accessor, InlineChatController.get(editor), this.desc.id);
-		}
-
-		let options: InlineChatRunOptions | undefined;
-		const arg = _args[0];
-		if (arg && InlineChatRunOptions.isInteractiveEditorOptions(arg)) {
-			options = arg;
-		}
-		InlineChatController.get(editor)?.run({ ...options });
-	}
-}
+import { Action2 } from 'vs/platform/actions/common/actions';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 
 export class HoldToSpeak extends AbstractInlineChatAction {
 
@@ -79,17 +37,23 @@ export class HoldToSpeak extends AbstractInlineChatAction {
 	}
 
 	override runInlineChatCommand(accessor: ServicesAccessor, ctrl: InlineChatController, editor: ICodeEditor, ...args: any[]): void {
-		holdForSpeech(accessor, ctrl, this.desc.id);
+		holdForSpeech(accessor, ctrl, this);
 	}
 }
 
-function holdForSpeech(accessor: ServicesAccessor, ctrl: InlineChatController | null, commandId: string): void {
+function holdForSpeech(accessor: ServicesAccessor, ctrl: InlineChatController, action: Action2): void {
+
+	const configService = accessor.get(IConfigurationService);
+	const speechService = accessor.get(ISpeechService);
 	const keybindingService = accessor.get(IKeybindingService);
 	const commandService = accessor.get(ICommandService);
-	if (!ctrl) {
+
+	// enabled or possible?
+	if (!configService.getValue<boolean>(InlineChatConfigKeys.HoldToSpeech || !speechService.hasSpeechProvider)) {
 		return;
 	}
-	const holdMode = keybindingService.enableKeybindingHoldMode(commandId);
+
+	const holdMode = keybindingService.enableKeybindingHoldMode(action.desc.id);
 	if (!holdMode) {
 		return;
 	}
@@ -109,3 +73,6 @@ function holdForSpeech(accessor: ServicesAccessor, ctrl: InlineChatController | 
 		handle.dispose();
 	});
 }
+
+// make this accessible to the chat actions from the browser layer
+setHoldForSpeech(holdForSpeech);
